Numba for q matrix
